All Teams View for NBAVis
The all-teams view now works… mostly. Not sure exactly why or how the
heat players get loaded in or what can safely be removed so that they
aren’t.

One major change that had to be made was that the database now stores
minutes played as seconds played (sadly, for convention, still calling
the column min). This is basically because time is complicated and to
calculate sums of time requires it being in a summable form.